add resampleListPar
Create as many cells as the input list is long and execute them in parallel
(in the sense that each one has a separate state). At each tick the list with
the different states grows or shrinks depending on the size of the input list.

Similar to Yampa's [parC](https://hackage.haskell.org/package/Yampa-0.13.3/docs/FRP-Yampa-Switches.html#v:parC).